фикс ~3,5к варнингов
осталось ещё 464 предупреждения, с ними буду уже потом разбираться.

+++ src/Layers/xrRenderPC_R1/LightShadows.cpp
@@ -253,8 +253,8 @@ void CLightShadows::calculate	()
 			mCombineR.mul			(mProjectR,mView);
 			
 			// Select slot and set viewport
-			int		s_x			=	slot_id%slot_line;
-			int		s_y			=	slot_id/slot_line;
+			u32		s_x			=	u32(slot_id%slot_line);
+			u32		s_y			=	u32(slot_id/slot_line);
 			D3DVIEWPORT9 VP		=	{s_x*S_size,s_y*S_size,S_size,S_size,0,1 };
 			CHK_DX					(HW.pDevice->SetViewport(&VP));
